# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  trips.TransportMode:
    enum:
    - flight
    - train
    - bus
    - car_rental
    - ferry
    - taxi
    - other
    type: string
    x-enum-varnames:
    - TransportModeFlight
    - TransportModeTrain
    - TransportModeBus
    - TransportModeCarRental
    - TransportModeFerry
    - TransportModeTaxi
    - TransportModeOther
  trips.TransportSegment:
    properties:
      arrive_at:
        type: string
      arrive_to:
        example: Mumbai CSMT
        type: string
      arrive_tz:
        example: Asia/Kolkata
        type: string
      booking_ref:
        example: PNR1234567
        type: string
      cost:
        example: 1500
        type: number
      created_at:
        type: string
      currency:
        example: INR
        type: string
      deep_link_url:
        example: https://www.irctc.co.in/nget/train-search
        type: string
      depart_at:
        type: string
      depart_from:
        example: New Delhi Railway Station
        type: string
      depart_tz:
        example: Asia/Kolkata
        type: string
      from_hop_id:
        type: string
      id:
        type: string
      mode:
        allOf:
        - $ref: '#/definitions/trips.TransportMode'
        example: train
      notes:
        type: string
      operator:
        example: IRCTC
        type: string
      seats:
        example: S4 42, S4 43
        type: string
      to_hop_id:
        type: string
      trip_plan:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
      updated_at:
        type: string
    type: object
  trips.TransportSegmentRequest:
    properties:
      arrive_at:
        example: "2026-09-01T22:00:00+05:30"
        type: string
      arrive_to:
        example: Mumbai CSMT
        type: string
      arrive_tz:
        example: Asia/Kolkata
        type: string
      booking_ref:
        example: PNR1234567
        type: string
      cost:
        example: 1500
        type: number
      currency:
        example: INR
        type: string
      deep_link_url:
        type: string
      depart_at:
        example: "2026-09-01T06:00:00+05:30"
        type: string
      depart_from:
        example: New Delhi Railway Station
        type: string
      depart_tz:
        example: Asia/Kolkata
        type: string
      from_hop_id:
        type: string
      mode:
        allOf:
        - $ref: '#/definitions/trips.TransportMode'
        example: train
      notes:
        type: string
      operator:
        example: IRCTC
        type: string
      seats:
        example: S4 42, S4 43
        type: string
      to_hop_id:
        type: string
    required:
    - mode
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API for managing trip plans, hops, and stays
  title: Trip Planner API
  version: "1.0"
paths:
...
  /transport/{id}:
    delete:
      parameters:
      - description: Segment ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: deleted
          schema:
            additionalProperties:
              type: boolean
            type: object
        "404":
          description: Not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Soft-delete a transport segment
      tags:
      - transport
    get:
      parameters:
      - description: Segment ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/trips.TransportSegment'
        "404":
          description: Not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get a transport segment by ID
      tags:
      - transport
    put:
      consumes:
      - application/json
      parameters:
      - description: Segment ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated fields
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/trips.TransportSegmentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/trips.TransportSegment'
        "404":
          description: Not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a transport segment
      tags:
      - transport
...
  /trip/{id}/transport:
    get:
      description: Returns all non-deleted transport segments for a trip plan, ordered
        by departure time.
      parameters:
      - description: Trip Plan ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: segments
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List transport segments for a trip
      tags:
      - transport
    post:
      consumes:
      - application/json
      description: Creates a new flight, train, bus, or other transport leg.
      parameters:
      - description: Trip Plan ID
        in: path
        name: id
        required: true
        type: string
      - description: Segment details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/trips.TransportSegmentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/trips.TransportSegment'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Trip not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Add a transport segment to a trip
      tags:
      - transport
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"